+++ modules/tinymce/src/core/test/ts/browser/content/EditorContentTest.ts
@@ -269,6 +269,25 @@ describe('browser.tinymce.core.content.EditorContentTest', () => {
           const tree = editor.getContent({ format: 'tree' });
           assert.equal(toHtml(tree), '<p>replaced</p>', 'Should be replaced html');
         });
+
+        const INITIAL_CONTENT = '<p>initial</p>';
+        const NEW_CONTENT = '<p>new content</p>';
+        const MANIPULATED_CONTENT = '<p>manipulated</p>';
+        Arr.each([
+          [ 'setContent', MANIPULATED_CONTENT ],
+          [ 'insertContent', MANIPULATED_CONTENT + '\n' + INITIAL_CONTENT ]
+        ] as const, ([ action, result ]) => {
+          it(`TINY-9143: Can manipulate content in "BeforeSetContent" callback when called from "${action}" function`, () => {
+            const editor = hook.editor();
+            editor.setContent(INITIAL_CONTENT);
+            editor.once('BeforeSetContent', (e) => {
+              assert.equal(e.content, NEW_CONTENT);
+              e.content = MANIPULATED_CONTENT;
+            });
+            editor[action](NEW_CONTENT);
+            TinyAssertions.assertContent(editor, result);
+          });
+        });
       });
     }
   );